requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
google-generativeai>=0.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
